Rename utils modules to logic
