--- server_django/project/messenger/utils/message_helpers.py
@@ -1,2 +1,0 @@
-def get_last_message_id(messages):
-    return messages.last().id
